Some refactor of handling enum conversions.
